Extract ObjectDetailView to a separate file - extract ObjectDetailWrapper.styled.tsx
