+++ frontend/src/metabase/visualizations/components/ObjectDetail/ObjectDetailWrapper.tsx
@@ -1,10 +1,8 @@
 import { useState } from "react";
-import type { ObjectDetailProps } from "metabase/visualizations/components/ObjectDetail/types";
-import {
-  PaginationFooter,
-  RootModal,
-} from "metabase/visualizations/components/ObjectDetail/ObjectDetail.styled";
-import { ObjectDetailView } from "metabase/visualizations/components/ObjectDetail/ObjectDetail";
+
+import { ObjectDetailView } from "./ObjectDetailView";
+import { PaginationFooter, RootModal } from "./ObjectDetailWrapper.styled";
+import type { ObjectDetailProps } from "./types";
 
 export function ObjectDetailWrapper({
   question,
